backend/requirements.txt
fastapi==0.115.0
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-docker>=7.0.0
GitPython>=3.1.40
prometheus-client>=0.19.0
